Refactor delete into a reviews controller. Code cleanup
