+++ src/main/frontend/components/PizzaStyleShow.js
@@ -1,15 +1,16 @@
 import React, {useEffect, useState} from "react"
-import { useLocation } from "react-router-dom"
+import {useLocation} from "react-router-dom"
 import ReviewTile from "./ReviewTile";
 import PizzaStyleTile from "./PizzaStyleTile";
-import {jsonGet, jsonPut, jsonPost, jsonDelete} from "./jsonFetch";
+import {jsonGet, jsonDelete} from "../public/js/jsonFetch";
 
 const PizzaStyleShow = props => {
   const [pizzaStyle, setPizzaStyle] = useState({reviews: []})
   let location = useLocation()
 
   const fetchPizzaStyle = async () => {
-    const respBody = await jsonGet(`/api/v1/pizza-styles/${props.match.params.id}`)
+    const respBody = await jsonGet(
+        `/api/v1/pizza-styles/${props.match.params.id}`)
     setPizzaStyle(respBody.pizzaStyle);
   }
 
@@ -17,18 +18,14 @@ const PizzaStyleShow = props => {
     fetchPizzaStyle()
   }, [location.pathname])
 
-  const callBack = (data) => {
-    fetchPizzaStyle();
+  const deleteReview = async (url) => {
+    await jsonDelete(url, fetchPizzaStyle);
   }
 
-  const deleteReview = async (url) => {
-    await jsonDelete(url, callBack);
-  }
   const onClick = event => {
     if (event.target.value.includes("delete")) {
       let id = event.target.value.split(":")[1];
-      deleteReview(`/api/v1/pizza-styles/delete/${id}`)
-
+      deleteReview(`/api/v1/reviews/${id}`)
     } else if (event.target.value.includes("edit")) {
       let id = event.target.value.split(":")[1];
       /***** TODO: Hit endPoint to redirect to prepopluated form ***
@@ -39,30 +36,30 @@ const PizzaStyleShow = props => {
 
   const reviewTiles = pizzaStyle.reviews.map(review => {
     return (
-        <div key={review.id}>
-          <ReviewTile
-              review={review}
-          />
-          <div className="small button-group">
-            <button type="button" value={`edit:${review.id}`} onClick={onClick}
-                    className="success button">Edit
-            </button>
-            <button type="button" value={`delete:${review.id}`}
-                    onClick={onClick} className="alert button">Delete
-            </button>
-          </div>
+      <div key={review.id}>
+        <ReviewTile
+            review={review}
+        />
+        <div className="small button-group">
+          <button type="button" value={`edit:${review.id}`} onClick={onClick}
+                  className="success button">Edit
+          </button>
+          <button type="button" value={`delete:${review.id}`}
+                  onClick={onClick} className="alert button">Delete
+          </button>
         </div>
+      </div>
     )
   })
 
   return (
-      <div>
-        <PizzaStyleTile
-            key={pizzaStyle.id}
-            pizzaStyle={pizzaStyle}
-        />
-        {reviewTiles}
-      </div>
+    <div>
+      <PizzaStyleTile
+          key={pizzaStyle.id}
+          pizzaStyle={pizzaStyle}
+      />
+      {reviewTiles}
+    </div>
   )
 }
 export default PizzaStyleShow
