+++ src/main/frontend/package.json
@@ -46,7 +46,7 @@
     "json-loader": "^0.5.4",
     "json2csv": "^4.4.1",
     "lodash": "^4.17.11",
-    "node-sass": "^6.0.0",
+    "node-sass": "^6.0.1",
     "pg": "^8.3.0",
     "react": "~16.8.0",
     "react-dom": "~16.8.0",
