+++ src/main/frontend/public/js/jsonFetch.js
@@ -75,19 +75,14 @@ const jsonDelete = async (url, callback) => {
       headers: new Headers({
         'Content-Type': 'application/json'
       }),
-      //body: JSON.stringify(payload)
     })
     if (!response.ok) {
       const errorMessage = `${response.status} (${response.statusText})`;
       const error = new Error(errorMessage);
       throw (error);
     }
-    //const body = await response.json();
-    //The model has been updated. We are done. Handle
-    //concat of data.strKey to existing state be handled by
     //client of this function, which will invoke a rendering
     callback();
-
   } catch (err) {
     console.error(`Error in fetch: ${err.status} (${err.message})`);
   }
