+++ src/main/java/com/launchacademy/reviews/controllers/PizzaStylesApiV1Controller.java
@@ -3,18 +3,17 @@ package com.launchacademy.reviews.controllers;
 import com.launchacademy.reviews.exceptionHandlers.CustomError;
 import com.launchacademy.reviews.models.PizzaStyle;
 import com.launchacademy.reviews.services.PizzaStyleService;
-import com.launchacademy.reviews.services.ReviewService;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -23,15 +22,12 @@ import org.springframework.web.bind.annotation.RestController;
 public class PizzaStylesApiV1Controller {
 
   private PizzaStyleService pizzaStyleService;
-  private ReviewService reviewService;
   private CustomError customError;
 
   @Autowired
   public PizzaStylesApiV1Controller(PizzaStyleService pizzaStyleService,
-      ReviewService reviewService,
       CustomError customError) {
     this.pizzaStyleService = pizzaStyleService;
-    this.reviewService = reviewService;
     this.customError = customError;
   }
 
@@ -74,10 +70,4 @@ public class PizzaStylesApiV1Controller {
     return map;
   }
 
-  @DeleteMapping("/delete/{id}")
-  public void deleteById(@PathVariable Integer id) {
-    System.out.println("deleteById( " + id + " )");
-    reviewService.deleteById(id);
-    System.out.println("post deleteById( " + id + " )");
-  }
 }
