+++ src/main/java/com/launchacademy/reviews/services/PizzaStyleService.java
@@ -37,10 +37,6 @@ public class PizzaStyleService {
     return repository.findById(id);
   }
 
-
-  public void deleteById(Integer id) {
-    repository.deleteById(id);
-  }
   public List<PizzaStyle> findByNameIgnoreCase(String name) {
     return repository.findByNameIgnoreCase(name);
   }
